chore: 🤖 create relative date helper
Create a relative date helper that outputs "time" ago format.

✅ Closes: ICU-5847
