--- addons/core/addon/components/relative-datetime-live/index.hbs
@@ -1,1 +1,0 @@
-{{this.formatted}}
