+++ addons/core/addon/components/toolbar-refresher/index.hbs
@@ -1,6 +1,6 @@
 {{#if this.lastRefreshed}}
   <Rose::Toolbar::Info>
-    <RelativeDatetimeLive @date={{this.lastRefreshed}} />
+    {{relative-datetime-live this.lastRefreshed}}
   </Rose::Toolbar::Info>
 {{/if}}
