--- addons/core/app/components/relative-datetime-live.js
@@ -1,1 +1,0 @@
-export { default } from 'core/components/relative-datetime-live';
