+++ addons/core/app/helpers/relative-datetime-live.js
@@ -1,0 +1,1 @@
+export { default } from 'core/helpers/relative-datetime-live';
