+++ addons/core/tests/dummy/app/templates/application.hbs
@@ -1,3 +1,2 @@
 <h2 id="title">Welcome to Ember</h2>
-
-<RelativeDatetimeLive @date={{this.now}} />
+<p>{{relative-datetime-live this.now}}</p>
